packaging

+++ clean.ps1
@@ -1,5 +1,12 @@
+param(
+    [switch][Alias("c")]$cleanCode
+)
+
 Remove-Item ".\*\obj\" -Force -Recurse
 Remove-Item ".\*\bin\" -Force -Recurse
-Get-ChildItem ".\*.tt" -Recurse | ForEach-Object {
-    Remove-Item $([io.path]::ChangeExtension($_.FullName, "cs")) -Force -ErrorAction Ignore
+
+if ($cleanCode) {
+    Get-ChildItem ".\*.tt" -Recurse | ForEach-Object {
+        Remove-Item $([io.path]::ChangeExtension($_.FullName, "cs")) -Force -ErrorAction Ignore
+    }
 }
